restructure test dir
