+++ tests/unit/components/text2cypher/validation/utils/test_cypher_extractors.py
@@ -1,28 +1,9 @@
-from ps_genai_agents.components.text2cypher.validation._utils import (
-    get_node_label_pattern,
-    get_node_pattern,
-    get_property_pattern,
+from ps_genai_agents.components.text2cypher.validation.utils.cypher_extractors import (
+    _extract_nodes_and_properties_from_cypher_statement,
+    _extract_relationships_and_properties_from_cypher_statement,
     parse_labels_or_types,
     process_match_clause_property_ids,
 )
-
-
-def test_get_property_pattern(prop_1: str) -> None: ...
-
-
-def test_get_node_pattern(node_1: str) -> None: ...
-
-
-def test_get_node_label_pattern() -> None:
-    pat = get_node_label_pattern()
-    assert pat.findall("(n:Node)")[0] == "Node"
-    assert pat.findall("(m {id:'001'})") == []
-    assert pat.findall("(nodeA:Node {id:'001'})")[0] == "Node "
-    assert pat.findall("(node_1:N)")[0] == "N"
-    assert pat.findall("(node_1:Node{id:'001'})")[0] == "Node"
-    assert pat.findall("(nodeA:`Node one` {id:'001'})")[0] == "Node one"
-    assert pat.findall("(node_1:Node|NodeB{id:'001'})")[0] == "Node|NodeB"
-    assert pat.findall("(node_1:Node&NodeB {id:'001'})")[0] == "Node&NodeB "
 
 
 def test_process_match_clause_property_ids_single(
@@ -101,3 +82,36 @@ def test_parse_labels_or_types_single_exclamation() -> None:
     res = parse_labels_or_types("!NodeA")
 
     assert len(res) == 0
+
+
+def test_extract_nodes_and_properties_from_cypher_statement_1(
+    cypher_statement_1: str,
+) -> None:
+    ents = _extract_nodes_and_properties_from_cypher_statement(cypher_statement_1)
+
+    answer = [
+        {"labels": "Node"},
+        {"labels": None, "property_name": "id", "property_value": "001"},
+    ]
+
+    no_labels = [x for x in answer if x.get("labels", "") is None]
+    assert len(ents) == len(answer)
+    assert no_labels[0].get("property_name") == "id"
+    assert no_labels[0].get("property_value") == "001"
+
+
+def test_extract_relationships_and_properties_from_cypher_statement_1(
+    cypher_statement_1: str,
+) -> None:
+    ents = _extract_relationships_and_properties_from_cypher_statement(
+        cypher_statement_1
+    )
+
+    answer = [
+        {"rel_type": "RELATIONSHIP", "property_name": None, "property_value": None}
+    ]
+
+    assert len(ents) == len(answer)
+    assert answer[0].get("rel_type") == "RELATIONSHIP"
+    assert answer[0].get("property_name", "wrong") is None
+    assert answer[0].get("property_value", "wrong") is None
